sourceCode/requirements.txt
astor==0.8.1
attrdict==2.0.1
babel==2.17.0
bce-python-sdk==0.9.35
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==2.1.1
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cssselect==1.3.0
cssutils==2.11.1
cycler==0.12.1
Cython==3.1.1
decorator==5.2.1
et_xmlfile==2.0.0
fire==0.7.0
Flask==3.1.1
flask-babel==4.0.0
fonttools==4.58.0
future==1.0.0
idna==3.10
imageio==2.37.0
imgaug==0.4.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
lazy_loader==0.4
lmdb==1.6.2
lxml==5.4.0
MarkupSafe
matplotlib==3.6.2
more-itertools==10.7.0
networkx==3.4.2
numpy==1.23.5
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opt-einsum==3.3.0
packaging==25.0
paddle-bfloat==0.1.7
paddleocr==2.6.1.3
paddlepaddle==2.4.2
pandas==2.2.3
pdf2docx==0.5.8
Pillow==9.4.0
premailer==3.10.0
protobuf==3.20.0
psutil==7.0.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pycryptodome==3.23.0
PyMuPDF==1.20.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.2
PyYAML==6.0
RapidFuzz==3.13.0
rarfile==4.2
requests>=2.31.0
scikit-image==0.24.0
scipy==1.9.3
seaborn==0.12.2
shapely==2.1.1
six==1.17.0
soupsieve==2.7
termcolor==3.1.0
tifffile==2025.5.10
torch==1.13.1
torchvision==0.14.1
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.0
ultralytics-thop==2.0.14
urllib3
visualdl==2.5.3
Werkzeug==3.1.3
gradio==3.50.2